refactored conversations to use a service
it works!!! well, the stuff i got working before now works again the
same way, but with a different back end.

next step is to redo characters and lines of dialogue to use services
as well.

weird note: from my research it seems you have to define the service in
EVERY view that uses it? this seems wild and maybe untrue but that
seems to be how it works nonetheless? i will keep looking into it.