new basic layout
making things look more like our paper prototype. the characters view
is now a list in a sidebar, and there's an area next to it for editing.
theres nothing to edit yet.

also it's really hackily styled right now and i need to fix it but i'm
really rusty re: web design apparently...........